requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
gunicorn
python-multipart

# ── Data validation ────────────────────────────────────────────────────────
pydantic[email]
pydantic-settings

# ── Database (MongoDB async) ───────────────────────────────────────────────
motor[srv]

# ── Environment ────────────────────────────────────────────────────────────
python-dotenv

# ── Authentication & Security ─────────────────────────────────────────────
python-jose[cryptography]
passlib[argon2]
firebase-admin
# HTTP rate limiting (Redis-backed, shared across workers). See
# app/rate_limit.py and app/middleware/*.
slowapi
# Redis client — already a transitive dep of celery[redis] but we depend
# on it directly for the rate limiter backend.
limits[redis]

# ── Cloudflare R2 (S3-compatible API, async via aioboto3) ─────────────────────
aioboto3

# ── HTTP client (async + sync) ────────────────────────────────────────────
httpx
requests

# ── Email (async SMTP) ────────────────────────────────────────────────────
aiosmtplib

# ── OTP / 2FA ─────────────────────────────────────────────────────────────
pyotp
qrcode[pil]

# ── AI Image Generation ───────────────────────────────────────────────────
# google-genai  → Vertex nano-banana-2 & nano-banana-pro providers
google-genai
Pillow

# ── Agentic photoshoot pipeline (LangGraph multi-agent orchestration) ─────
# The provider fallback chain (kie_nb2 → vertex_nb2 → vertex_nbpro →
# evolink_nb2) is implemented as a LangGraph StateGraph in
# app/services/photoshoot_service.py.
langgraph

# ── Photoshoot / Image processing ─────────────────────────────────────────
numpy
opencv-python-headless

# ── Queue management (Celery + Redis) ─────────────────────────────────────
celery[redis]
redis
